Add archive page and year pages generation
Following the same approach as with the index page generation, I'm using
the same template but populate it with links.

Every year page is the same as an index page but with posts for that year
only, so I could re-use a lot of code.

Tweaked the styles a bit as well.